Fixed the requested issues
